Initial stab at getting inputs & data types exporting/importing correctly - before, the shared input ids could get shared across teams, or skills within a team - there was also a potential to have params with user-defined data types not work correctly after import
I'd like to do a cleanup pass on this, as I think this code can be structured in a way to better avoid these sorts of problems. One issue is that we are trying to use the same JSON structures for our editors and for import/export. I'm sure that seemed like a great idea at the time.