Fix for the problem where a selected input is being ignored.
One thing that can happen is that you pass a value through to a data type (say a string). It might not be valid to start, but once the data type fetched options, if it's an exact match, it becomes valid. However, if the data type returns other options and you try to select one of them, the selection gets ignored because the data type doesn't think it has anything else to collect.

A better fix for this might be to pass through the intended param from the input choice selection, but this is a start